Move a few grid header out of details
