+++ core/include/detray/builders/grid_factory.hpp
@@ -14,9 +14,9 @@
 #include "detray/geometry/concepts.hpp"
 #include "detray/geometry/mask.hpp"
 #include "detray/geometry/shapes.hpp"
-#include "detray/utils/grid/detail/axis.hpp"
+#include "detray/utils/grid/axis.hpp"
+#include "detray/utils/grid/concepts.hpp"
 #include "detray/utils/grid/detail/axis_helpers.hpp"
-#include "detray/utils/grid/detail/concepts.hpp"
 #include "detray/utils/grid/grid.hpp"
 #include "detray/utils/grid/grid_collection.hpp"
 #include "detray/utils/grid/populators.hpp"
