+++ core/include/detray/core/detail/type_traits.hpp
@@ -12,7 +12,7 @@
 #include "detray/material/material_rod.hpp"
 #include "detray/material/material_slab.hpp"
 #include "detray/navigation/accelerators/concepts.hpp"
-#include "detray/utils/grid/detail/concepts.hpp"
+#include "detray/utils/grid/concepts.hpp"
 #include "detray/utils/type_registry.hpp"
 
 // System include(s)
