+++ core/include/detray/utils/grid/detail/axis_helpers.hpp
@@ -11,7 +11,7 @@
 #include "detray/definitions/algebra.hpp"
 #include "detray/definitions/grid_axis.hpp"
 #include "detray/geometry/coordinates/coordinates.hpp"
-#include "detray/utils/grid/detail/axis.hpp"
+#include "detray/utils/grid/axis.hpp"
 #include "detray/utils/grid/detail/axis_binning.hpp"
 #include "detray/utils/grid/detail/axis_bounds.hpp"
 #include "detray/utils/type_list.hpp"
