+++ core/include/detray/utils/grid/detail/bin_view.hpp
@@ -10,8 +10,8 @@
 // Project include(s)
 #include "detray/definitions/containers.hpp"
 #include "detray/definitions/detail/qualifiers.hpp"
+#include "detray/utils/grid/concepts.hpp"
 #include "detray/utils/grid/detail/axis_bounds.hpp"
-#include "detray/utils/grid/detail/concepts.hpp"
 #include "detray/utils/ranges.hpp"
 
 // System include(s)
