+++ core/include/detray/utils/grid/grid.hpp
@@ -10,7 +10,7 @@
 // Project include(s).
 #include "detray/core/detail/container_views.hpp"
 #include "detray/definitions/detail/qualifiers.hpp"
-#include "detray/utils/grid/detail/axis.hpp"
+#include "detray/utils/grid/axis.hpp"
 #include "detray/utils/grid/detail/axis_helpers.hpp"
 #include "detray/utils/grid/detail/bin_storage.hpp"
 #include "detray/utils/grid/populators.hpp"
