+++ io/include/detray/io/backend/detail/grid_writer.hpp
@@ -13,7 +13,7 @@
 #include "detray/io/backend/detail/type_info.hpp"
 #include "detray/io/frontend/payloads.hpp"
 #include "detray/utils/detector_statistics.hpp"
-#include "detray/utils/grid/detail/concepts.hpp"
+#include "detray/utils/grid/concepts.hpp"
 #include "detray/utils/grid/grid.hpp"
 #include "detray/utils/type_list.hpp"
 #include "detray/utils/type_traits.hpp"
